Moved hotspot popover to inner placeholder
